% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{splintr}
\alias{splintr}
\title{Generate a basis matrix for natural cubic splines with interpretable intercepts}
\usage{
splintr(..., centre = 0)
}
\arguments{
\item{...}{Parameters passed to \code{\link[splines]{ns}}.}

\item{centre}{The \code{x}-value at which all spline terms should be zero.}
}
\value{
A matrix of dimension \code{length(x)} by \code{df}, like \code{\link[splines]{ns}}.
  Attributes are that correspond to the arguments to \code{\link[splines]{ns}} with an additional
  \code{centre} attribute for use by \code{predict.splintr()}.
}
\description{
Generate a basis matrix for natural cubic splines with interpretable intercepts
}
